Renamed folder to "example"
